Setup app stucture and create basic rover icon component Co-authored-by: Shane Misra <shane.misra@gmail.com> Co-authored-by: Bea Ordonez <beaordonez25@gmail.com>
