--- src/components/App/App.css
@@ -1,61 +1,0 @@
-/* .App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-float infinite 3s ease-in-out;
-  }
-}
-
-@keyframes App-logo-float {
-  0% {
-    transform: translateY(0);
-  }
-  50% {
-    transform: translateY(10px);
-  }
-  100% {
-    transform: translateY(0px);
-  }
-} */
-.App {
-  display:flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-  width: 100vw;
-  background-image: linear-gradient(132deg, #d28c24 0%, #ad4b00 100%);
-}
-
-.welcomeText {
-  font-size: 70px;
-  margin: 25px;
-  padding: 19px;
-  color: aliceblue;
-  border: solid 1px;
-  border-radius: 28px;
-}
-
-.friendsOfJohnnyFive {
-  display: flex;
-  padding: 30px;
-  margin: 12px;
-}
-
-.basicRobot {
-  height: 150px;
-  width: 150px;
-  padding: 12px;
-  border: solid 1px;
-  border-radius: 90px;
-  opacity: .9;
-  background-color: aliceblue;
-  box-shadow: rgba(0, 0, 0, 0.25) 0px 54px 55px, rgba(0, 0, 0, 0.12) 0px -12px 30px, rgba(0, 0, 0, 0.12) 0px 4px 6px, rgba(0, 0, 0, 0.17) 0px 12px 13px, rgba(0, 0, 0, 0.09) 0px -3px 5px;
-}
